+++ local/templates/yugavto.expert.theme/components/bitrix/news.list/main.cards.cars/result_modifier.php
@@ -1,7 +1,6 @@
 <?php
-    $url = 'https://apps.yug-avto.ru/API/get/cis/count/used/?token=34b5ac8b71018c0bc7e5c050ed90b243&!dealership=1489&city='.YApp::setCityCookie();
+    $url = 'https://apps.yug-avto.ru/API/get/cis/count/used/?token=34b5ac8b71018c0bc7e5c050ed90b243&city='.YApp::setCityCookie();
     $arResult['COUNTS']['PASS'] = file_get_contents($url);
-    $url = 'https://apps.yug-avto.ru/API/get/cis/count/used/?token=34b5ac8b71018c0bc7e5c050ed90b243&dealership=1489&city='.YApp::setCityCookie();
-    $arResult['COUNTS']['COMM'] = file_get_contents($url);
+    $arResult['COUNTS']['COMM'] = $arResult['COUNTS']['PASS'];
 
     // YApp::sp($arResult['ITEMS'], true);